# docs/context/engineering/api-contracts/v1/openapi.yaml
openapi: 3.1.0
info:
  title: InComedy Server API
  version: "1.0.0"
  description: |
    Current backend API contract for auth, identity foundation, organizer workspace foundation,
    and service health endpoints.
servers:
  - url: https://incomedy.ru
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/session/me:
    get:
      tags: [Auth]
      summary: Validate access token and return current session user
      operationId: getSessionMe
      parameters:
        - in: header
          name: Authorization
          required: true
          description: Bearer access token.
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Access token is valid and user exists
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionMeResponse"
        "401":
          description: Missing/invalid token or unknown user
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server failure
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke current user session tokens up to now
      operationId: postAuthLogout
      parameters:
        - in: header
          name: Authorization
          required: true
          description: Bearer access token.
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session revoked successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
        "401":
          description: Missing/invalid token
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server failure
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new session tokens
      operationId: postAuthRefresh
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Optional client-provided UUID request id for tracing.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          description: Invalid payload or invalid refresh token format
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds route size limit
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing/invalid refresh token or session revoked
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected server failure
          headers:
            X-Request-ID:
              description: Server trace identifier
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    SessionMeResponse:
      type: object
      additionalProperties: false
      required:
        - provider
        - user
      properties:
        provider:
          type: string
          enum: [password, phone, vk, telegram, google, apple]
        user:
          $ref: "#/components/schemas/SessionUser"
...
    RefreshRequest:
      type: object
      additionalProperties: false
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      additionalProperties: false
      required:
        - provider
        - access_token
        - refresh_token
        - expires_in
        - user
      properties:
        provider:
          type: string
          enum: [password, phone, vk, telegram, google, apple]
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
          example: 3600
        user:
          $ref: "#/components/schemas/SessionUser"
    SessionUser:
      type: object
      additionalProperties: false
      required:
        - id
        - display_name
        - roles
        - linked_providers
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          example: John Doe
        username:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [audience, comedian, organizer]
        active_role:
          type: string
          nullable: true
          enum: [audience, comedian, organizer]
        linked_providers:
          type: array
          items:
            type: string
            enum: [password, phone, vk, telegram, google, apple]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: auth_failed
        message:
          type: string
          example: Authentication failed